Tests/LibWeb/Text/expected/css/modal-dialog-display-toggle-lays-out.txt

modal: true
computed display: block
has box: false
width: 0
